More conversions to Kotlin

+++ radar-commons-android/src/main/java/org/radarcns/util/BackedObjectQueue.java
@@ -115,9 +115,9 @@ public class BackedObjectQueue<S, T> implements Closeable {
      * @throws IOException if the element could not be read or deserialized
      * @throws IllegalStateException if the element could not be read
      */
-    public List<T> peek(int n, int sizeLimit) throws IOException {
+    public List<T> peek(int n, long sizeLimit) throws IOException {
         Iterator<InputStream> iter = queueFile.iterator();
-        int curSize = 0;
+        long curSize = 0;
         List<T> results = new ArrayList<>(n);
         for (int i = 0; i < n && iter.hasNext(); i++) {
             try (InputStream in = iter.next()) {
